elsai-ocr-extractors/pyproject.toml
[project]
name = "elsai-ocr-extractors"
version = "0.1.0"
description = ""
authors = [
    {name = "abhinivesh0305",email = "abhinivesh.s@optisolbusiness.com"}
]
requires-python = ">=3.9,<3.14"
dependencies = [
    "boto3 (>=1.38.32,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "amazon-textract-textractor (>=1.9.2,<2.0.0)",
    "azure-ai-documentintelligence (>=1.0.2,<2.0.0)",
    "azure-core (>=1.34.0,<2.0.0)",
    "azure-cognitiveservices-vision-computervision (>=0.9.1,<0.10.0)",
    "openai (>=1.84.0,<2.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "llama-parse (>=0.6.30,<0.7.0)"
 
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
